+++ CryptocurrencyExchange/Services/WalletService.cs
@@ -2,7 +2,6 @@
 using CryptocurrencyExchange.Models;
 using CryptocurrencyExchange.Services.Interfaces;
 using CryptocurrencyExchange.Services.Wallet;
-using Microsoft.EntityFrameworkCore;
 
 namespace CryptocurrencyExchange.Services
 {
@@ -34,7 +33,7 @@ namespace CryptocurrencyExchange.Services
             var usdt = await _walletItemRepository.GetAsync(userId, "usdt")
             ?? throw new WalletItemNotFoundException("USDT in wallet not found");
 
-            if ((decimal)usdt.Amount < usd)
+            if (usdt.Amount < usd)
                 throw new InsufficientFundsException();
 
             var coin = await _walletItemRepository.GetAsync(userId, coinSymbol)
@@ -48,7 +47,7 @@ namespace CryptocurrencyExchange.Services
         }
 
 
-        public async Task<double> GetCoinAmountAsync(int userId, string symbol)
+        public async Task<decimal> GetCoinAmountAsync(int userId, string symbol)
         {
             var walletItem = await _walletItemRepository.GetAsync(userId, symbol);
             return walletItem?.Amount ?? 0;
@@ -80,7 +79,7 @@ namespace CryptocurrencyExchange.Services
         }
 
 
-        public async Task SellAsync(int userId, string coinSymbol, double amount)
+        public async Task SellAsync(int userId, string coinSymbol, decimal amount)
         {
             coinSymbol = coinSymbol.ToLower();
 
@@ -93,8 +92,8 @@ namespace CryptocurrencyExchange.Services
             if (coin.Amount < amount)
                 throw new InsufficientFundsException($"Not enough balance in {coinSymbol.ToUpper()}");
 
-            var coinPrice = (decimal)await _marketService.GetPrice(coinSymbol);
-            _walletDomainService.Sell(usdt, coin, (decimal)amount, coinPrice);
+            var coinPrice = await _marketService.GetPrice(coinSymbol);
+            _walletDomainService.Sell(usdt, coin, amount, coinPrice);
             await _unitOfWork.CommitAsync();
         }
     }
